Tokenization/ResoucePivots: - Replacing all specific Locales/Themes resource code with a generic ResourcePivot system that allows an x number of pivots,. - Adding DPI to WebGrease instead of abusing Themes for it - Added an option to run Resource/Tokanization after parsing of css - Added Multi-Threading to MinifyCssActivity, works best with after parse token/resources. - Created FileSetBase to be root for both Css/Js filesets with common functionality. CSS parser: - Support for hash() in url() - Enabled %TOKENVAUE% as valid REPLACEMENTTOKEN for term, property and string values. - Enabled ___TOKEN___ as token replacement code for selectors. - Updated some invalid comments - Updated tests, and added test for all new scenario's - Removed -wg-dpi because is now handled by DPI support. General: - Several small fixes and updates to caching/multi-threading. - Added Safe.Lock to allow for locking with a timeout. - Added a ParallelForEach method to the context. - Added file lock to ensure only only webgrease per cache folder at a time. - Updates for codeflow comments.

+++ WebGrease/WebGrease/Css/ImageAssemblyAnalysis/PropertyModel/Background.cs
@@ -94,7 +94,7 @@ namespace WebGrease.Css.ImageAssemblyAnalysis.PropertyModel
         /// <param name="updatedY">The updated y</param>
         /// <param name="webGreaseBackgroundDpi">The webgrease background dpi to use</param>
         /// <returns>The new declaration node with updated values</returns>
-        internal DeclarationNode UpdateBackgroundNode(string updatedUrl, int? updatedX, int? updatedY, double webGreaseBackgroundDpi)
+        internal DeclarationNode UpdateBackgroundNode(string updatedUrl, int? updatedX, int? updatedY, float webGreaseBackgroundDpi)
         {
             var isUrlUpdated = false;
             var isXUpdated = false;
@@ -118,7 +118,6 @@ namespace WebGrease.Css.ImageAssemblyAnalysis.PropertyModel
                         continue;
                     }
                 }
-
 
                 // Try updating X
                 if (!isXUpdated)
